Removed all unnecessary from the code (Imports, methods, tests) removed the myCalc construct from tests and created a special class for it.
