+++ calculator/src/main/java/calculator/MainCalculator.java
@@ -1,6 +1,5 @@
 package calculator;
 
-import java.util.InputMismatchException;
 import java.util.Scanner;
 import javax.naming.OperationNotSupportedException;
